+++ R/keepCells.R
@@ -44,7 +44,7 @@ keepCells <- function(cnr, keep) {
     cnr[["cells"]] <- keep
 
     if(!is.null(cnr[["vdj.cells"]])) {
-        vk <- cnr[["vdj.cells"]] %in% keep
+        vk <- names(cnr[["vdj.cells"]]) %in% keep
         cnr[["vdj.cells"]] <- cnr[["vdj.cells"]][vk]
     }